restructure, manually add suspense boundary
