--- src/app/best/loading.tsx
@@ -1,10 +1,0 @@
-import { Feed } from "~/components/Feed";
-import { FeedPageLayout } from "~/components/FeedPagesLayout";
-
-export default function Loading() {
-  return (
-    <FeedPageLayout>
-      <Feed />
-    </FeedPageLayout>
-  );
-}
