+++ src/app/explore/collected/page.tsx
@@ -1,6 +1,6 @@
 import { ExplorePublicationType, ExplorePublicationsOrderByType } from "@lens-protocol/client";
-import { FeedPageLayout } from "~/components/FeedPagesLayout";
 import { InfiniteScroll } from "~/components/InfiniteScroll";
+import { FeedPageLayout } from "~/components/layout/FeedLayout";
 import { lensItemToPost } from "~/components/post/Post";
 import { getLensClient } from "~/utils/getLensClient";
